requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
jinja2
aiofiles
httpx
python-multipart
websockets
python-dotenv
huggingface_hub
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
joblib>=1.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake-useragent>=1.4.0
asyncio-throttle>=1.0.0
pip install googletrans==4.0.0rc1
